Remove timestamp from jenkins build (#13)
Per slack discussion, the date timestamp is not super helpful when built with jenkins, and because of changes necessary for Java 11, tools were going to have issues with this functionality. Local builds still have the timestamp, and jenkins builds are still stable and increasing because describe still have a commit count since last tag.